+++ src/Components/ChatWrapper/index.js
@@ -1,0 +1,3 @@
+import { ChatWrapper } from "./Chats";
+
+export { ChatWrapper };
