+++ src/Router/Router.sass
@@ -12,43 +12,6 @@ a
   margin: 0 auto
   display: flex
 
-.Chat-wrapper
-  width: 100%
-  max-width: 360px
-  display: flex
-  flex-direction: column
-  position: relative
-
-.Add-chat-open-modal
-  --size: 54px
-  --offset: 1.25rem
-  border-radius: 50%
-  height: var(--size)
-  width: var(--size)
-  line-height: var(--size)
-  position: absolute
-  transform: translateZ(0)
-  --translateY: calc(54px + var(--offset))
-  bottom: var(--offset)
-  right: var(--offset)
-  background-color: #3390ec
-  transition: all .3s ease
-  font-size: 35px
-  text-align: center
-  color: white
-  cursor: pointer
-  z-index: 99
-  box-shadow: 0 0 5px rgba(0, 0, 0, 0.63)
-  &:hover
-    background-color: darken(#3390ec, 20%)
-  &:disabled
-    opacity: 50%
-    cursor: inherit
-    &:hover
-      background-color: #3390ec
-  &:active
-    background-color: darken(#3390ec, 30%)
-
